requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
asttokens==3.0.0
attrs==24.3.0
autopep8==2.3.2
azure-common==1.1.28
azure-ml==0.0.1
blinker==1.9.0
cachetools==5.5.0
certifi==2025.1.31
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
narwhals==1.18.4
nest-asyncio==1.6.0
numpy==2.2.0
opencensus-context==0.1.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pycodestyle==2.12.1
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.8.0
pylint==3.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.19.0
scipy==1.14.1
setuptools==77.0.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
streamlit==1.41.1
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tomlkit==0.13.2
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.34.0
watchdog==4.0.1
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
openai>=0.27.0
transformers>=4.28.0
torch>=2.0.0
accelerate>=0.29.3
sentencepiece>=0.2.0
